server/requirements.txt
fastapi == 0.100.0
uvicorn == 0.22.0
hupper == 1.12
pymysql == 1.1.0
python-dotenv == 1.0.0
sqlalchemy == 2.0.19
alembic == 1.11.1
passlib == 1.7.4
pydantic == 2.1.1
bcrypt == 4.0.1